add group optimization method for stitch registration; add c-zarr reader/writer; update zarr reader/writer for MIP, inplace decon

+++ microscopeDataProcessing/zarr/writezarr.m
@@ -2,6 +2,9 @@ function writezarr(data, filepath, varargin)
 % wrapper for zarr writer 
 % 
 % Author: Xiongtao Ruan (01/25/2022)
+% 
+% xruan (05/23/2022): add support for bounding box write; also add
+% parallelWriteZarr as default method
 
 
 ip = inputParser;
@@ -11,12 +14,14 @@ ip.addRequired('filepath', @(x) ischar(x));
 ip.addParameter('blockSize', [500, 500, 500], @isnumeric);
 ip.addParameter('expand2dDim', true, @islogical); % expand the z dimension for 2d data
 ip.addParameter('groupWrite', true, @islogical);
+ip.addParameter('bbox', [], @isnumeric);
 ip.parse(data, filepath, varargin{:});
 
 pr = ip.Results;
 expand2dDim = pr.expand2dDim;
 blockSize = pr.blockSize;
 groupWrite = pr.groupWrite;
+bbox = pr.bbox;
 
 dtype = class(data);
 sz = size(data);
@@ -31,17 +36,30 @@ if ismatrix(data)
 end
 blockSize = min(sz, blockSize);
 % overwrite the zarr file
-if exist(filepath, 'dir')
+if exist(filepath, 'dir') && isempty(bbox)
     rmdir(filepath, 's');
 end
-bim = blockedImage(filepath, sz, blockSize, init_val, "Adapter", ZarrAdapter, 'Mode', 'w');
-% for data greater than 2GB, use multiprocessing
-if ~ispc && prod(sz) * 2 / 1024^3 > 2
-    bim.Adapter.setData(data);
-else
-    bim.Adapter.setRegion(ones(1, numel(bim.Size)), bim.Size, data)
+
+try 
+    if ismatrix(data)
+        error('No support for 2d data for now!')
+    end
+    if isempty(bbox)
+        parallelWriteZarr(filepath, data, 1, blockSize);
+    else
+        parallelWriteZarr(filepath, data, 1, bbox);
+    end
+catch ME
+    disp(ME);
+    bim = blockedImage(filepath, sz, blockSize, init_val, "Adapter", ZarrAdapter, 'Mode', 'w');
+    % for data greater than 2GB, use multiprocessing
+    if ~ispc && prod(sz) * 2 / 1024^3 > 2 && ~ismatrix(data) 
+        bim.Adapter.setData(data);
+    else
+        bim.Adapter.setRegion(ones(1, numel(bim.Size)), bim.Size, data)
+    end
+    bim.Adapter.close();
 end
-bim.Adapter.close();
 
 if groupWrite
     try
